+++ R/module-chartControls.R
@@ -10,9 +10,9 @@
 #'
 #' @importFrom shinyWidgets dropdown
 #' @importFrom htmltools tags tagList
-#' @importFrom shiny icon
+#' @importFrom shiny icon checkboxInput
 #'
-chartControlsUI <- function(id, insert_code = FALSE) {
+chartControlsUI <- function(id, insert_code = FALSE, disable_filters = FALSE) {
 
   # Namespace
   ns <- NS(id)
@@ -34,7 +34,8 @@ chartControlsUI <- function(id, insert_code = FALSE) {
       status = "default btn-controls"
     ),
     dropdown(
-      controls_params(ns), controls_appearance(ns),
+      controls_params(ns), 
+      controls_appearance(ns),
       style = "default",
       label = "Plot options",
       up = TRUE, 
@@ -42,19 +43,18 @@ chartControlsUI <- function(id, insert_code = FALSE) {
       icon = icon("gears"), 
       status = "default btn-controls"
     ),
-    dropdown(
-      tags$div(
-        style = "max-height: 400px; overflow-y: scroll; overflow-x: hidden;", #  padding-left: 10px;
-        filterDF_UI(id = ns("filter-data"))
-      ),
-      style = "default", 
-      label = "Data", 
-      up = TRUE, 
-      icon = icon("filter"),
-      right = TRUE, 
-      inputId = "filterdrop",
-      status = "default btn-controls"
-    ),
+    if (!isTRUE(disable_filters)) {
+      dropdown(
+        filterDF_UI(id = ns("filter-data")),
+        style = "default", 
+        label = "Data", 
+        up = TRUE, 
+        icon = icon("filter"),
+        right = TRUE, 
+        inputId = "filterdrop",
+        status = "default btn-controls"
+      )
+    },
     dropdown(
       controls_code(ns, insert_code = insert_code), 
       style = "default", 
@@ -68,9 +68,6 @@ chartControlsUI <- function(id, insert_code = FALSE) {
     tags$script("$('.sw-dropdown').addClass('btn-group-charter');"),
     tags$script(HTML("$('.sw-dropdown > .btn').addClass('btn-charter');")),
     tags$script("$('#sw-content-filterdrop').click(function (e) {e.stopPropagation();});"),
-    tags$script("$('#sw-content-filterdrop').css('min-width', '350px');"),
-    tags$script("$('#sw-content-codedrop').css('min-width', '350px');"),
-    tags$script("$('#sw-content-paramsdrop').css('min-width', '330px');"),
     useShinyUtils()
   )
 }
@@ -134,7 +131,7 @@ chartControlsServer <- function(input, output, session,
         gg <- ggplot_rv$ggobj$plot
         ppt <- officer::read_pptx()
         ppt <- officer::add_slide(ppt, layout = "Title and Content", master = "Office Theme")
-        ppt <- try(rvg::ph_with_vg(ppt, ggobj = gg, type = "body"), silent = TRUE)
+        ppt <- try(officer::ph_with(ppt, rvg::dml(ggobj = gg), location = officer::ph_location_type(type = "body")), silent = TRUE)
         if ("try-error" %in% class(ppt)) {
           shiny::showNotification(ui = "Export to PowerPoint failed...", type = "error")
         } else {
@@ -157,18 +154,22 @@ chartControlsServer <- function(input, output, session,
     context <- rstudioapi::getSourceEditorContext()
     code <- ggplot_rv$code
     code <- stri_replace_all(str = code, replacement = "+\n", fixed = "+")
-    if (input$insert_code == 1) {
-      code <- paste("library(ggplot2)", code, sep = "\n\n")
-    }
     if (!is.null(output_filter$code$expr)) {
       code_dplyr <- deparse(output_filter$code$dplyr, width.cutoff = 80L)
       code_dplyr <- paste(code_dplyr, collapse = "\n")
       nm_dat <- data_name()
-      code_dplyr <- paste(nm_dat, code_dplyr, sep = " <- ")
       code_dplyr <- stri_replace_all(str = code_dplyr, replacement = "%>%\n", fixed = "%>%")
-      code <- paste(code_dplyr, code, sep = "\n\n")
+      code <- stri_replace_all(str = code, replacement = " ggplot()", fixed = sprintf("ggplot(%s)", nm_dat))
+      code <- paste(code_dplyr, code, sep = " %>%\n")
+      if (input$insert_code == 1) {
+        code <- paste("library(dplyr)\nlibrary(ggplot2)", code, sep = "\n\n")
+      }
+    } else {
+      if (input$insert_code == 1) {
+        code <- paste("library(ggplot2)", code, sep = "\n\n")
+      }
     }
-    rstudioapi::insertText(text = paste0("\n", code), id = context$id)
+    rstudioapi::insertText(text = paste0("\n", code, "\n"), id = context$id)
   })
   
   output$code <- renderUI({
@@ -178,9 +179,9 @@ chartControlsServer <- function(input, output, session,
       code_dplyr <- deparse(output_filter$code$dplyr, width.cutoff = 80L)
       code_dplyr <- paste(code_dplyr, collapse = "\n")
       nm_dat <- data_name()
-      code_dplyr <- paste(nm_dat, code_dplyr, sep = " <- ")
       code_dplyr <- stri_replace_all(str = code_dplyr, replacement = "%>%\n", fixed = "%>%")
-      code <- paste(code_dplyr, code, sep = "\n\n")
+      code <- stri_replace_all(str = code, replacement = " ggplot()", fixed = sprintf("ggplot(%s)", nm_dat))
+      code <- paste(code_dplyr, code, sep = " %>%\n")
     }
     htmltools::tagList(
       rCodeContainer(id = ns("codeggplot"), code)
@@ -289,7 +290,16 @@ chartControlsServer <- function(input, output, session,
     data_name = data_name
   )
 
-  outin <- reactiveValues(inputs = NULL, export_ppt = NULL, export_png = NULL)
+  outin <- reactiveValues(
+    inputs = NULL, 
+    export_ppt = NULL, 
+    export_png = NULL
+  )
+  
+  observeEvent(data_table(), {
+    outin$data <- data_table()
+    outin$code <- reactiveValues(expr = NULL, dplyr = NULL)
+  })
 
   observeEvent({
     all_inputs <- reactiveValuesToList(input)
@@ -429,35 +439,11 @@ controls_labs <- function(ns) {
 #' @importFrom shiny icon
 #' @importFrom htmltools tagList tags
 #' @importFrom shinyWidgets colorSelectorInput pickerInput radioGroupButtons spectrumInput
-#' @importFrom ggplot2 theme_bw theme_classic theme_dark theme_gray theme_grey 
-#'  theme_light theme_linedraw theme_minimal theme_void
-#' @importFrom ggthemes theme_base theme_calc theme_economist theme_economist_white 
-#'  theme_excel theme_few theme_fivethirtyeight theme_foundation theme_gdocs theme_hc 
-#'  theme_igray theme_map theme_pander theme_par theme_solarized theme_solarized_2 
-#'  theme_solid theme_stata theme_tufte theme_wsj
-#' @importFrom hrbrthemes theme_ft_rc theme_ipsum theme_ipsum_ps theme_ipsum_rc
-#'  theme_ipsum_tw theme_modern_rc
 controls_appearance <- function(ns) {
 
-  themes <- list(
-    ggplot2 = list(
-      "bw", "classic", "dark", "gray",
-      "light", "linedraw", "minimal",
-      "void"
-    ),
-    hrbrthemes = c(
-      "ft_rc", "ipsum", "ipsum_ps", "ipsum_rc", "ipsum_tw", "modern_rc"
-    ),
-    ggthemes = list(
-      "base", "calc", "economist", "economist_white",
-      "excel", "few", "fivethirtyeight", "foundation",
-      "gdocs", "hc", "igray", "map", "pander",
-      "par", "solarized", "solarized_2", "solid",
-      "stata", "tufte", "wsj"
-    )
-  )
-
-  cols <- colors_palettes()
+  themes <- get_themes()
+  cols <- get_colors()
+  pals <- get_palettes()
 
   tagList(
     tags$div(
@@ -465,25 +451,17 @@ controls_appearance <- function(ns) {
       spectrumInput(
         inputId = ns("fill_color"),
         label = "Choose a color:",
-        choices = c(list(c("#0C4C8A", "#EF562D")), unname(cols$choices_colors)), 
+        choices = unname(cols), 
         width = "100%"
       )
     ),
     tags$div(
       id = ns("controls-palette"), style = "display: none;",
-      tags$style(".bootstrap-select .dropdown-menu li a span.text {width: 100%;}"),
-      pickerInput(
-        inputId = ns("palette"),
-        label = "Choose a palette:",
-        choices = cols$colors_pal,
-        selected = "ggplot2", 
-        width = "100%",
-        choicesOpt = list(
-          content = sprintf(
-            "<div style='width:100%%;border-radius:4px; padding: 2px;background:%s;color:%s'>%s</div>",
-            unname(cols$background_pals), cols$colortext_pals, names(cols$background_pals)
-          )
-        )
+      palettePicker(
+        inputId = ns("palette"), 
+        label = "Choose a palette:", 
+        choices = pals$choices,
+        textColor = pals$textColor
       )
     ),
     pickerInput(
@@ -692,94 +670,39 @@ controls_code <- function(ns, insert_code = FALSE) {
 
 
 
+# Get list of themes
+get_themes <- function() {
+  themes <- getOption("esquisse.themes")
+  if (is.function(themes))
+    themes <- themes()
+  if (!is.list(themes)) {
+    stop("Option 'esquisse.themes' must be a list", call. = FALSE)
+  }
+  themes
+}
 
+# Get list of palettes
+get_palettes <- function() {
+  pals <- getOption("esquisse.palettes")
+  if (is.function(pals))
+    pals <- pals()
+  if (!is.list(pals)) {
+    stop("Option 'esquisse.palettes' must be a list with at least one slot : 'choices'", call. = FALSE)
+  }
+  if (is.null(pals$textColor))
+    pals$textColor <- "white"
+  pals
+}
 
-
-
-
-#' Colors for picker
-#'
-#' @noRd
-#'
-#' @importFrom RColorBrewer brewer.pal brewer.pal.info
-#' @importFrom scales hue_pal viridis_pal
-#' @importFrom hrbrthemes ipsum_pal ft_pal
-colors_palettes <- function() {
-  ### colors
-  # For spectrum pre-defined colors
-  choices_colors <- list(
-    "viridis" = col2Hex(viridis_pal(option = "viridis")(10)),
-    "magma" = col2Hex(viridis_pal(option = "magma")(10)),
-    "inferno" = col2Hex(viridis_pal(option = "inferno")(10)),
-    "plasma" = col2Hex(viridis_pal(option = "plasma")(10)),
-    "cividis" = col2Hex(viridis_pal(option = "cividis")(10))
-    ,
-    "ipsum" = ipsum_pal()(9),
-    "ft" = ft_pal()(9)
-    ,
-    "Blues" = get_brewer_pal(name = "Blues"),
-    "Greens" = get_brewer_pal(name = "Greens"),
-    "Reds" = get_brewer_pal(name = "Reds"),
-    "Oranges" = get_brewer_pal(name = "Oranges"),
-    "Purples" = get_brewer_pal(name = "Purples"),
-    "Greys" = get_brewer_pal(name = "Greys"),
-    "Dark2" = get_brewer_pal(name = "Dark2"),
-    "Set1" = get_brewer_pal(name = "Set1"),
-    "Paired" = get_brewer_pal(name = "Paired")
-  )
-
-  # For palette picker
-  colors_pal <- list(
-    Diverging = list(
-      "BrBG", "PiYG", "PRGn", "PuOr", "RdBu", 
-      "RdGy", "RdYlBu", "RdYlGn", "Spectral"
-    ), 
-    Qualitative = list(
-      "Accent", "Dark2", "Paired", "Pastel1", "Pastel2", "Set1", 
-      "Set2", "Set3"
-    ),
-    Sequential = list(
-      "Blues", "BuGn", "BuPu", 
-      "GnBu", "Greens", "Greys", "Oranges", "OrRd", "PuBu", "PuBuGn", 
-      "PuRd", "Purples", "RdPu", "Reds", "YlGn", "YlGnBu", "YlOrBr", 
-      "YlOrRd"
-    )
-  )
-  background_pals <- sapply(unlist(colors_pal, use.names = FALSE), get_brewer_pal)
-  # add ggplot2 hue & viridis
-  background_pals <- c(
-    list("ggplot2" = scales::hue_pal()(9)),
-    list(
-      "viridis" = col2Hex(viridis_pal(option = "viridis")(10)),
-      "magma" = col2Hex(viridis_pal(option = "magma")(10)),
-      "inferno" = col2Hex(viridis_pal(option = "inferno")(10)),
-      "plasma" = col2Hex(viridis_pal(option = "plasma")(10)),
-      "cividis" = col2Hex(viridis_pal(option = "cividis")(10))
-    ),
-    list(
-      "ipsum" = ipsum_pal()(9),
-      "ft" = ft_pal()(9)
-    ),
-    background_pals
-  )
-  background_pals <- unlist(lapply(X = background_pals, FUN = linear_gradient))
-  colortext_pals <- rep(c("white", "black", "black"), times = sapply(colors_pal, length))
-  colortext_pals <- c("white", rep("white", 5), rep("white", 2), colortext_pals) # ggplot2 + viridis + hrbrthemes
-
-  # add ggplot2 hue & viridis
-  colors_pal <- c(
-    list("Default" = list("ggplot2")),
-    list("Viridis" = list("viridis", "magma", "inferno", "plasma", "cividis")),
-    list("hrbrthemes" = list("ipsum", "ft")),
-    colors_pal
-  )
-
-  list(
-    colors_pal = colors_pal,
-    background_pals = background_pals,
-    colortext_pals = colortext_pals,
-    choices_colors = choices_colors
-  )
+# Get list of colors (spectrum)
+get_colors <- function() {
+  cols <- getOption("esquisse.colors")
+  if (is.function(cols))
+    cols <- cols()
+  if (!is.list(cols)) {
+    stop("Option 'esquisse.colors' must be a list", call. = FALSE)
+  }
+  cols
 }